refactor(core): improve code quality and remove dead code - Encapsulate Server fields - Refactor ByteBufferPool - Remove unused loop in KademliaPeerLookupJob - Clean up unused code in various classes

+++ src/main/java/im/redpanda/core/LocalSettings.java
@@ -62,7 +62,9 @@ public class LocalSettings implements Serializable {
 
       File file = new File(Settings.SAVE_DIR + "/localSettings" + port + ".dat");
 
-      file.createNewFile();
+      if (!file.createNewFile() && !file.exists()) {
+        throw new IOException("Could not create file " + file);
+      }
       try (FileOutputStream fileOutputStream = new FileOutputStream(file)) {
         try (ObjectOutputStream objectOutputStream = new ObjectOutputStream(fileOutputStream)) {
           objectOutputStream.writeObject(this);
